extract shootclient logic from crb state to separate package
