+++ internal/controller/runtime/fsm/runtime_fsm_patch2_shoot_test.go
@@ -3,6 +3,7 @@ package fsm
 import (
 	"context"
 	imv1 "github.com/kyma-project/infrastructure-manager/api/v1"
+	imv1_client "github.com/kyma-project/infrastructure-manager/internal/controller/runtime/fsm/client"
 	fsm_testing "github.com/kyma-project/infrastructure-manager/internal/controller/runtime/fsm/testing"
 	. "github.com/onsi/ginkgo/v2" //nolint:revive
 	. "github.com/onsi/gomega"    //nolint:revive
@@ -37,7 +38,7 @@ var _ = Describe("KIM patch2", func() {
 				ShootClient: fakeClient,
 				Client:      fakeClient,
 			}}
-			GetShootClientPatch = func(
+			imv1_client.GetShootClientPatch = func(
 				_ context.Context,
 				_ client.Client,
 				_ imv1.Runtime) (client.Client, error) {
